backend: rename docs folder to swagger
